change component
