+++ mpfar/pages/index.tsx
@@ -4,32 +4,24 @@ import Layout from "../components/Layout"
 import Post, { PostProps } from "../components/Post"
 import Profile, { ProfileProps } from "../components/Profile"
 import thesis, { ThesisProps } from "../components/Thesis"
+import Search from "../components/Search"
 import prisma from '../lib/prisma'
 import { PrismaClient } from "@prisma/client"
 import Thesis from "../components/Thesis"
 
 export const getStaticProps: GetStaticProps = async () => {
   const prisma = new PrismaClient()
-  const feed = await prisma.post.findMany({
-    where: { published: true },
-    include: {
-      author: {
-        select: { name: true },
-      },
-    },
-  });
   const data = await prisma.profile.findMany();
   const profile = JSON.parse(JSON.stringify(data));
   const data_t = await prisma.thesis.findMany();
   const thesisfile = JSON.parse(JSON.stringify(data_t));
   return {
-    props: { feed, profile,thesisfile},
+    props: { profile, thesisfile},
     revalidate: 10,
   };
 };
 
 type Props = {
-  feed: PostProps[],
   profile: ProfileProps[],
   thesisfile:ThesisProps[]
 }
@@ -40,31 +32,28 @@ const Blog: React.FC<Props> = (props) => {
     console.log(props.thesisfile)
 
     const loginUser:ProfileProps = props.profile.find((user) => user.id === 'aaa')
+    const loginUserThesis:ThesisProps[] = props.thesisfile.filter((thesis) => {
+      return thesis.firstAuthorId === loginUser.id
+    })
 
   return (
     <Layout>
       <div className="page">
         <h1>タイムライン</h1>
         <main>
-          {props.feed.map((post) => (
-            <div key={post.id} className="post">
-              <Post post={post} />
-            </div>
-          ))}
-        </main>
-        <main>
             <div className="post">
             <Profile profile={loginUser}/>
             </div>
         </main>
         <main>
-          {props.thesisfile.map((post3) => (
+          {loginUserThesis.map((post3) => (
             <div key={post3.id} className="post">
             <Thesis thesis={post3} />
             </div>
           ))}
         </main>
       </div>
+      <Search />
       <style jsx>{`
         .post {
           background: white;
